Split cmd for readability sake
Try to ease discovery/readability by splitting cmd (cobra/viper
setup) in various files. The proeminent cmd/execute.go should
hopefuly make the flow clear:
```
main() -> execute -> run.Run() -> {http, recorder, observer, events, git}
```
