+++ README.md
@@ -58,22 +58,22 @@ Available Commands:
   version     Print the version number
 
 Flags:
-  -s, --api-server string            Kubernetes api-server url
-  -c, --config string                Configuration file (default "/etc/katafygio/katafygio.yaml")
-  -d, --dry-run                      Dry-run mode: don't store anything.
-  -m, --dump-only                    Dump mode: dump everything and exit
-  -x, --exclude-kind stringSlice     Ressource kind to exclude. Eg. 'deployment'
-  -y, --exclude-object stringSlice   Object to exclude. Eg. 'configmap:kube-system/kube-dns'
-  -l, --filter string                Label filter. Select only objects matching the label.
-  -g, --git-url string               Git repository URL
-  -p, --healthcheck-port int         Port for answering healthchecks on /health url
-  -h, --help                         help for katafygio
-  -k, --kube-config string           Kubernetes config path
-  -e, --local-dir string             Where to dump yaml files (default "./kubernetes-backup")
-  -v, --log-level string             Log level (default "info")
-  -o, --log-output string            Log output (default "stderr")
-  -r, --log-server string            Log server (if using syslog)
-  -i, --resync-interval int          Resync interval in seconds (0 to disable) (default 300)
+  -s, --api-server string        Kubernetes api-server url
+  -c, --config string            Configuration file (default "/etc/katafygio/katafygio.yaml")
+  -d, --dry-run                  Dry-run mode: don't store anything
+  -m, --dump-only                Dump mode: dump everything once and exit
+  -x, --exclude-kind strings     Ressource kind to exclude. Eg. 'deployment'
+  -y, --exclude-object strings   Object to exclude. Eg. 'configmap:kube-system/kube-dns'
+  -l, --filter string            Label filter. Select only objects matching the label
+  -g, --git-url string           Git repository URL
+  -p, --healthcheck-port int     Port for answering healthchecks on /health url
+  -h, --help                     help for katafygio
+  -k, --kube-config string       Kubernetes config path
+  -e, --local-dir string         Where to dump yaml files (default "./kubernetes-backup")
+  -v, --log-level string         Log level (default "info")
+  -o, --log-output string        Log output (default "stderr")
+  -r, --log-server string        Log server (if using syslog)
+  -i, --resync-interval int      Full resync interval in seconds (0 to disable) (default 900)
 ```
 
 ## Config file and env variables
